// OMKARAUTOAPP/src/app/models/challanreleasemodel.ts
export class ChallanreleaseModel {
    chReleaseId: string = "";
    chYear: string = "";
    challanBranch: string = "";
    challanNo: string = "";
    challanId: string = "";
    releaseForPmt: string = "";
    year: string = "";
    branch: string = "";
    loggedInUser: string = "";
    createdBy: string = "";
    createdDate: string = "";
    modifiedBy: string = "";
    modifiedDate: string = "";
     
  }